+++ app/controllers/folio/console/links_controller.rb
@@ -24,17 +24,9 @@ class Folio::Console::LinksController < Folio::Console::BaseController
   private
 
     def page_links
-      if Rails.application.config.folio_pages_ancestry
-        {
-          Folio::Page => Proc.new do |page|
-            nested_page_path(page, add_parents: true)
-          end
-        }
-      else
-        {
-          Folio::Page => Proc.new { |page| main_app.url_for([page, only_path: true]) }
-        }
-      end
+      {
+        Folio::Page => Proc.new { |page| main_app.url_for([page, only_path: true]) }
+      }
     end
 
     def additional_links