+++ CHANGELOG.md
@@ -3,14 +3,15 @@ All notable changes to this project will be documented in this file.
 
 ## 2019-02-15
 ### Changed
-- Translations are not enabled by default, one has to set `Rails.application.config.folio_pages_translations = true`
-- Pages ancestry is not enabled by default, one has to set `Rails.application.config.folio_pages_ancestry = true`
+- Translations are not enabled by default, set `Rails.application.config.folio_pages_translations = true` to enable.
+- Pages ancestry is not enabled by default, set `Rails.application.config.folio_pages_ancestry = true` to enable.
 - Renamed `folio_nodes` to `folio_pages`, change the STI default to `Folio::Page`.
 - Page slugs now must be unique, no scoping.
 
 ### Removed
 - Removed `Folio::Node`, `Folio::Category` and `Folio::NodeTranslation`.
-- Removed `content` from `Folio::Page`
+- Removed `content` from `Folio::Page`.
+- Removed `nested_page_path` helper, use `url_for` instead.
 
 ## 2019-02-04
 ### Changed